The activity xtext plugin generates the activity model by interacting with the engine server

+++ engine/src/main/java/pt/ist/socialsoftware/blendedworkflow/domain/NotCondition.java
@@ -125,7 +125,8 @@ public class NotCondition extends NotCondition_Base {
 
     @Override
     public void delete() {
-        getCondition().delete();
+        if (getCondition() != null)
+            getCondition().delete();
         super.delete();
     }
